add icon & size
