+++ packages/client/src/feature/session/register/ui/RegisterPasswordInput/RegisterPasswordInput.tsx
@@ -1,29 +1,36 @@
 import { FC, useState } from 'react'
 
-// TODO: feature/cfg-23 icon вынести в отдельную папкку
-import VisibilityIcon from '@mui/icons-material/Visibility'
-import VisibilityOffIcon from '@mui/icons-material/VisibilityOff'
-
 import BaseTextField from '@/shared/components/ui/BaseTextField'
 
 import BaseIconButton from '@/shared/components/ui/BaseIconButton'
 
+import { IconVisibility, IconVisibilityOff } from '@/shared/components/icons/iconsMui'
+
 import { InputProps } from '../../types'
 
-const RegisterPasswordInput: FC<InputProps> = ({ formData, handleChange, errors, name, label, required = true }) => {
+const RegisterPasswordInput: FC<InputProps> = ({
+  formData,
+  handleChange,
+  errors,
+  name,
+  label,
+  required = true,
+  size = 'small'
+}) => {
   const [showPassword, setShowPassword] = useState(false)
 
   const handleTogglePasswordVisibility = () => {
     setShowPassword(prevShowPassword => !prevShowPassword)
   }
 
-  const visibleIcon = showPassword ? <VisibilityIcon /> : <VisibilityOffIcon />
+  const visibleIcon = showPassword ? <IconVisibility /> : <IconVisibilityOff />
   const endIcon = {
     endAdornment: <BaseIconButton onClick={handleTogglePasswordVisibility}>{visibleIcon}</BaseIconButton>
   }
 
   return (
     <BaseTextField
+      size={size}
       required={required}
       type={showPassword ? 'text' : 'password'}
       label={label}
