+++ packages/client/src/shared/components/icons/iconsMui.ts
@@ -1,0 +1,7 @@
+import LocalPhoneIcon from '@mui/icons-material/LocalPhone'
+import VisibilityIcon from '@mui/icons-material/Visibility'
+import VisibilityOffIcon from '@mui/icons-material/VisibilityOff'
+
+export const IconLocalPhone = LocalPhoneIcon
+export const IconVisibility = VisibilityIcon
+export const IconVisibilityOff = VisibilityOffIcon
