+++ packages/client/src/widgets/session/register/ui/RegisterWidget/RegisterWidget.tsx
@@ -1,9 +1,15 @@
 import { FC } from 'react'
 
+import BaseContainer from '@/shared/components/ui/BaseContainer'
+
 import RegisterFeature from '@/feature/session/register'
 
 const RegisterWidget: FC = () => {
-  return <RegisterFeature />
+  return (
+    <BaseContainer maxWidth="sm">
+      <RegisterFeature />
+    </BaseContainer>
+  )
 }
 
 export default RegisterWidget
